% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMAlignR.R
\docType{package}
\name{BAMAlignR}
\alias{BAMAlignR}
\title{BAMAlignR package}
\description{
This package is designed to read in a BAM/SAM file, allow users to choose a specific read entry
and display how that read's sequence aligns to the reference genome using its CIGAR string.
}
\author{
Cristina Iudica \email{cristina.iudica@mail.polimi.it}
}
