% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMAlignR_functions.R
\name{CheckStrand}
\alias{CheckStrand}
\title{Check the strand orientation of the mapping read from the FLAG field of the BAM file}
\usage{
CheckStrand(read_flag)
}
\arguments{
\item{read_flag}{FLAG value associated to the read of interest}
}
\value{
A message indicating whether the read is mapped to the forward or reverse strand.
}
\description{
This function checks if the read is mapping on the forward strand or on the reverse strand by taking as input the FLAG value from the BAM file.
The 5th bit (16) of the FLAG field indicates whether the read is reverse complemented or not.
This bit is set (1) if the read is reverse complemented, and it is not set (0) if it is not reverse complemented.
It performs a bitwise AND operation ('&') between the flag field and the value 16 (00010000).
The strand orientation of the mapping read is printed out.
}
\examples{
CheckStrand(16) # Output: "The read is mapping to the reverse strand"
CheckStrand(0)  # Output: "The read is mapping to the forward strand"

}
