% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMAlignR_functions.R
\name{CigarAnalyzerReference}
\alias{CigarAnalyzerReference}
\title{Analyze the CIGAR string and print the aligned reference sequence}
\usage{
CigarAnalyzerReference(read_CIGAR, read_rname, length_aligned_seq, read_pos)
}
\arguments{
\item{read_CIGAR}{CIGAR string associated to the read of interest}

\item{read_rname}{String representing the reference sequence name to which the read is aligned}

\item{length_aligned_seq}{Integer representing the length of the aligned read}

\item{read_pos}{Integer representing the starting position of the mapping read on the reference sequence}
}
\value{
A string representing the reference sequence that aligns with the read
}
\description{
This function uses the \link{ExtractReferenceSequence} function to extract the reference sequence of interest, and then based on the CIGAR string, it adds a '-' into the sequence
if the CIGAR operation is I (i.e., insertion to the reference sequence), in such a way that the reference sequence correctly aligns with the read.
It uses the same logic as \link{CigarAnalyzerRead} function, but inverting insertions and deletions/skipped regions.
}
\examples{
read_CIGAR <- "3M2I5M"
read_seq_char <- "TCTAACGC"
read_rname <- 11
length_aligned_seq <- 10
read_pos <- 100
aligned_ref_seq <- CigarAnalyzerReference(read_CIGAR, read_rname, length_aligned_seq, read_pos)
cat("Aligned Reference Sequence:", aligned_ref_seq, "\n")

}
\references{
\url{https://samtools.github.io/hts-specs/SAMv1.pdf}
}
\seealso{
\code{\link{ExtractReferenceSequence}}
}
