% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMAlignR_functions.R
\name{ExtractReadInfoAndCheck}
\alias{ExtractReadInfoAndCheck}
\title{Extract information from a specific read and perform some checks}
\usage{
ExtractReadInfoAndCheck(bamPath, readID)
}
\arguments{
\item{bamPath}{Path to the BAM file}

\item{readID}{A character string representing the ID of the read to extract information for}
}
\value{
A list containing the following information about the read of interest:
\item{read_rname}{The reference sequence name to which the read is aligned.}
\item{read_pos}{The leftmost mapping position of the first matching base within the read.}
\item{read_CIGAR}{The CIGAR string representing the alignment pattern of the read.}
\item{read_seq_char}{The sequence of the read, stored as a character string.}
\item{read_flag}{The flag value associated with the read.}

#' @examples
bamPath <- system.file("extdata", "HG00096.chrom11.ILLUMINA.bwa.bam", package = "BAMAlignR")
readID <- "SRR062635.8528185"
readInfo <- ExtractReadInfoAndCheck(bamPath, readID)
}
\description{
This function takes as input a BAM file path and a specific read ID and it extracts some relevant information for that read,
i.e. the sequence (chromosome or contig) within the reference genome where the read is aligned (RNAME), the leftmost mapping position of the first matching base within the read (POS),
the CIGAR string, the actual read sequence, the FLAG value, the strand to which the read maps.
Additionally, it verifies that the input parameters are correct and that the BAM file contains the proper information about the read of interest.
}
\references{
\url{https://samtools.github.io/hts-specs/SAMv1.pdf}
}
