% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMAlignR_functions.R
\name{ExtractReferenceSequence}
\alias{ExtractReferenceSequence}
\title{Extract the reference sequence}
\usage{
ExtractReferenceSequence(read_rname, length_aligned_seq, read_pos, read_CIGAR)
}
\arguments{
\item{read_rname}{String representing the reference sequence name to which the read is aligned}

\item{length_aligned_seq}{Integer representing the length of the aligned read}

\item{read_pos}{Integer representing the starting position of the mapping read on the reference sequence}

\item{read_CIGAR}{CIGAR string associated to the read of interest}
}
\value{
A string containing the extracted reference sequence
}
\description{
This function takes as input the sequence within the reference genome where the read is aligned (i.e. RNAME in the BAM file, that is extracted in the function \link{ExtractReadInfoAndCheck}),
the length of the reference sequence you need to extract (i.e., the length of the aligned read, that is computed in the function \link{CigarAnalyzerRead}),
the position of the nucleotide in the reference genome from which the extracted sequence starts (i.e. POS in the BAM file, that is the adjusted with the function
\link{AdjustStartingPosition} and the CIGAR string.
From this information, the function extracts the reference sequence to which the read is aligned.
}
\examples{
read_rname <- 11
length_aligned_seq <- 100
read_pos <- 100
read_CIGAR <- "100M"
ref_sequence <- ExtractReferenceSequence(read_rname, length_aligned_seq, read_pos, read_CIGAR)

cat("Extracted Reference Sequence:", ref_sequence, "\n")

}
\seealso{
\code{\link{AdjustStartingPosition}}
}
