% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMAlignR_functions.R
\name{ShowAlignment}
\alias{ShowAlignment}
\title{Show the read alignment and the CIGAR string}
\usage{
ShowAlignment(bamPath, readID)
}
\arguments{
\item{bamPath}{Path to the BAM file}

\item{readID}{A character string representing the ID of the read}
}
\value{
A representation of the  precise alignment of the read to the reference genome according to the mapped read's CIGAR string
}
\description{
Core function of the package: it shows the precise alignment of the read to the reference genome according to the mapped
read's CIGAR string. It first extracts relevant information using the \link{ExtractReadInfoAndCheck} function. It then
utilizes the \link{CigarAnalyzerRead} and \link{CigarAnalyzerReference} functions to process the CIGAR string and generate
aligned sequences for both the read and the reference. The function then prints the CIGAR string, the aligned reference
sequence, and the aligned read sequence.
}
\examples{
bamPath <- system.file(
"extdata",
"HG00096.chrom11.ILLUMINA.bwa.bam",
package = "BAMAlignR"
)
readID <- "SRR062635.8528185"
ShowAlignment(bamPath, readID)

}
\seealso{
\code{\link{ExtractReadInfoAndCheck}}, \code{\link{CigarAnalyzerRead}}, \code{\link{CigarAnalyzerReference}}
}
